refactor: add messageClass / feat(whisper): socket
