+++ .dart_tool/package_config_subset
@@ -22,6 +22,10 @@ cross_file
 2.19
 file:///C:/Users/4_da_/AppData/Local/Pub/Cache/hosted/pub.dev/cross_file-0.3.3+5/
 file:///C:/Users/4_da_/AppData/Local/Pub/Cache/hosted/pub.dev/cross_file-0.3.3+5/lib/
+crypto
+2.19
+file:///C:/Users/4_da_/AppData/Local/Pub/Cache/hosted/pub.dev/crypto-3.0.3/
+file:///C:/Users/4_da_/AppData/Local/Pub/Cache/hosted/pub.dev/crypto-3.0.3/lib/
 cupertino_icons
 2.19
 file:///C:/Users/4_da_/AppData/Local/Pub/Cache/hosted/pub.dev/cupertino_icons-1.0.6/
@@ -158,6 +162,10 @@ vector_math
 2.14
 file:///C:/Users/4_da_/AppData/Local/Pub/Cache/hosted/pub.dev/vector_math-2.1.4/
 file:///C:/Users/4_da_/AppData/Local/Pub/Cache/hosted/pub.dev/vector_math-2.1.4/lib/
+web_socket_channel
+2.15
+file:///C:/Users/4_da_/AppData/Local/Pub/Cache/hosted/pub.dev/web_socket_channel-2.4.0/
+file:///C:/Users/4_da_/AppData/Local/Pub/Cache/hosted/pub.dev/web_socket_channel-2.4.0/lib/
 sky_engine
 3.0
 file:///C:/Users/4_da_/Downloads/flutter_windows_3.10.5-stable/flutter/bin/cache/pkg/sky_engine/
